# openapi.yaml
openapi: 3.0.3
info:
  title: ERP/POS API
  description: |
    Sistema ERP de punto de venta y gestión de almacén.
    
    ## Características principales
    - Autenticación JWT con roles
    - Gestión de productos con variantes
    - Control de inventario en tiempo real
    - Punto de venta con precios automáticos (mayoreo/menudeo)
    - Facturación SAT preparada
    - Multi-caja y multi-usuario
    
    ## Arquitectura
    - Backend: FastAPI + PostgreSQL
    - Frontend: React
    - Base de datos con lógica de negocio centralizada
    
  version: 1.0.0
  contact:
    name: Soporte Técnico
    email: soporte@erp.com
servers:
  - url: http://localhost:8000/api/v1
    description: Desarrollo local
  - url: https://api.erp.com/api/v1
    description: Producción
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenido del endpoint /auth/login.
        Incluye en el header: `Authorization: Bearer {token}`

  # ============================================
  # ESQUEMAS
  # ============================================
  schemas:
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: "STOCK_INSUFFICIENT"
        message:
          type: string
          example: "Stock insuficiente para completar la operación"
        details:
          type: object
          additionalProperties: true

    # --- Auth ---
...
    VarianteBarcodeResponse:
      type: object
      properties:
        variante_id:
          type: integer
          example: 3
        producto_id:
          type: integer
          example: 1
        nombre_producto:
          type: string
          example: "Playera Básica"
        nombre_variante:
          type: string
          example: "Playera Negra Talla M"
        codigo_barras:
          type: string
          example: "7501234567890"
        sku:
          type: string
          example: "PLAY-NEG-M"
        precio_menudeo:
          type: number
          format: decimal
          example: 600.00
        precio_mayoreo:
          type: number
          format: decimal
          example: 520.00
        umbral_mayoreo:
          type: integer
          example: 12
        stock_actual:
          type: integer
          example: 42
        activo:
          type: boolean
          example: true
    ProductoDetalle:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        descripcion:
          type: string
        categoria:
          type: string
        marca:
          type: string
        activo:
          type: boolean
        variantes:
          type: array
          items:
            $ref: '#/components/schemas/VarianteBarcodeResponse'

    # --- Ventas ---
    VentaCreateRequest:
      type: object
      required:
        - punto_venta_id
        - metodo_pago
      properties:
        punto_venta_id:
          type: integer
          example: 1
        cliente_nombre:
          type: string
          nullable: true
          example: "Juan Pérez"
        metodo_pago:
          type: string
          enum: [EFECTIVO, TARJETA, TRANSFERENCIA]
          example: "EFECTIVO"
    VentaCreateResponse:
      type: object
      properties:
        venta_id:
          type: integer
          example: 101
        folio:
          type: string
          example: "PV1-000101"
        fecha_creacion:
          type: string
          format: date-time
        estado:
          type: string
          enum: [ABIERTA, CERRADA, CANCELADA]
          example: "ABIERTA"
...
    VentaDetalle:
      type: object
      properties:
        venta_id:
          type: integer
        folio:
          type: string
        fecha_venta:
          type: string
          format: date-time
        cliente_nombre:
          type: string
          nullable: true
        metodo_pago:
          type: string
        estado:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              detalle_id:
                type: integer
              nombre_producto:
                type: string
              nombre_variante:
                type: string
              cantidad:
                type: integer
              precio_unitario:
                type: number
              subtotal:
                type: number
        subtotal:
          type: number
        descuento_total:
          type: number
        total:
          type: number
        usuario:
          type: string
        punto_venta:
          type: string
...
    MovimientoInventarioRequest:
      type: object
      required:
        - variante_id
        - tipo
        - cantidad
        - motivo
      properties:
        variante_id:
          type: integer
          example: 3
        tipo:
          type: string
          enum: [ENTRADA, SALIDA, AJUSTE]
          example: "ENTRADA"
        cantidad:
          type: integer
          minimum: 1
          example: 20
        motivo:
          type: string
          example: "Compra a proveedor XYZ"
        referencia:
          type: string
          nullable: true
          example: "OC-12345"
    StockVariante:
      type: object
      properties:
        variante_id:
          type: integer
        nombre_producto:
          type: string
        nombre_variante:
          type: string
        sku:
          type: string
        stock_actual:
          type: integer
        stock_minimo:
          type: integer
        estado_stock:
          type: string
          enum: [OK, BAJO, CRITICO]

    # --- Facturación ---
...
paths:
...
  /ventas:
    post:
      tags:
        - Ventas
      summary: Crear una nueva venta
      description: |
        Crea una venta nueva en estado ABIERTA.
        Esta es la primera operación del flujo POS.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VentaCreateRequest'
      responses:
        '201':
          description: Venta creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VentaCreateResponse'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - Ventas
      summary: Listar ventas
      description: Obtiene un listado de ventas con filtros opcionales
      security:
        - BearerAuth: []
      parameters:
        - name: estado
          in: query
          schema:
            type: string
            enum: [ABIERTA, CERRADA, CANCELADA]
        - name: punto_venta_id
          in: query
          schema:
            type: integer
        - name: fecha_desde
          in: query
          schema:
            type: string
            format: date
        - name: fecha_hasta
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Lista de ventas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VentaDetalle'
...
  /productos:
    get:
      tags:
        - Productos
      summary: Listar productos
      security:
        - BearerAuth: []
      parameters:
        - name: categoria
          in: query
          schema:
            type: string
        - name: buscar
          in: query
          schema:
            type: string
          description: Búsqueda por nombre o SKU
        - name: activo
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de productos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductoDetalle'
  /productos/{producto_id}:
    get:
      tags:
        - Productos
      summary: Obtener detalle de producto
      security:
        - BearerAuth: []
      parameters:
        - name: producto_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalle del producto con todas sus variantes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoDetalle'

  # ============================================
  # INVENTARIO
  # ============================================
  /inventario/movimientos:
    post:
      tags:
        - Inventario
      summary: Registrar movimiento de inventario
      description: |
        Registra una entrada, salida o ajuste de inventario.
        Las salidas por venta se registran automáticamente, esto es para:
        - Compras a proveedores (ENTRADA)
        - Mermas (SALIDA)
        - Ajustes (AJUSTE)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovimientoInventarioRequest'
      responses:
        '201':
          description: Movimiento registrado
        '409':
          description: Stock insuficiente para salida
  /inventario/stock:
    get:
      tags:
        - Inventario
      summary: Consultar stock de variantes
      security:
        - BearerAuth: []
      parameters:
        - name: estado
          in: query
          schema:
            type: string
            enum: [OK, BAJO, CRITICO]
          description: Filtrar por estado de stock
      responses:
        '200':
          description: Listado de stock
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockVariante'
  /inventario/stock/bajo:
    get:
      tags:
        - Inventario
      summary: Productos con stock bajo
      description: Devuelve variantes con stock <= stock_minimo
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Productos con stock bajo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockVariante'

  # ============================================
  # FACTURACIÓN
  # ============================================
...